# swagger-spec.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"Characters","description":"These endpoints are used to manage the Characters SQL repository."}],"paths":{"/characters":{"get":{"tags":["Characters"],"summary":"Get all the Marvel Character IDs","operationId":"getAllCharacterIdsUsingGET","produces":["*/*"],"responses":{"200":{"description":"Character IDs Populated","schema":{"$ref":"#/definitions/ResponseEntity"}},"204":{"description":"No Content","schema":{"$ref":"#/definitions/ResponseEntity"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/characters/seed":{"get":{"tags":["Characters"],"summary":"Update the H2 SQL Database with all of the Marvel Characters from the official API","operationId":"seedCharactersUsingGET","produces":["*/*"],"responses":{"200":{"description":"Database Successfully Seeded, please GET /characters or /characters/MARVEL_ID","schema":{"$ref":"#/definitions/ResponseEntity"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"511":{"description":"Google API Key Required","schema":{"$ref":"#/definitions/Error"}}},"deprecated":false}},"/characters/{characterId}":{"get":{"tags":["Characters"],"summary":"Get a Marvel Character's Name, Description and Image, and optionally translate the Description to another language (2-letter ISO-639-1 language code)","operationId":"getCharacterUsingGET","produces":["*/*"],"parameters":[{"name":"characterId","in":"path","description":"characterId","required":true,"type":"integer","format":"int32"},{"name":"language","in":"query","description":"language","required":false,"type":"string","default":"en","allowEmptyValue":false}],"responses":{"200":{"description":"Character Found","schema":{"$ref":"#/definitions/ResponseEntity"}},"400":{"description":"Incorrect Language Provided, please use ISO-639-1 language codes","schema":{"$ref":"#/definitions/Error"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Character Not Found","schema":{"$ref":"#/definitions/Error"}},"511":{"description":"Google API Key Required","schema":{"$ref":"#/definitions/Error"}}},"deprecated":false}}},"definitions":{"Error":{"type":"object","properties":{"cause":{"$ref":"#/definitions/Throwable"},"localizedMessage":{"type":"string"},"message":{"type":"string"},"stackTrace":{"type":"array","items":{"$ref":"#/definitions/StackTraceElement"}},"suppressed":{"type":"array","items":{"$ref":"#/definitions/Throwable"}}},"title":"Error"},"ResponseEntity":{"type":"object","properties":{"body":{"type":"object"},"statusCode":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"statusCodeValue":{"type":"integer","format":"int32"}},"title":"ResponseEntity"},"StackTraceElement":{"type":"object","properties":{"classLoaderName":{"type":"string"},"className":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"methodName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"nativeMethod":{"type":"boolean"}},"title":"StackTraceElement"},"Throwable":{"type":"object","properties":{"cause":{"$ref":"#/definitions/Throwable"},"localizedMessage":{"type":"string"},"message":{"type":"string"},"stackTrace":{"type":"array","items":{"$ref":"#/definitions/StackTraceElement"}},"suppressed":{"type":"array","items":{"$ref":"#/definitions/Throwable"}}},"title":"Throwable"}}}